requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pycryptodome==3.19.0
bcrypt==4.1.2
qrcode==7.4.2
Pillow==10.2.0
python-dotenv==1.0.0
